Refactor: Change structure of hooks directory
* Move previous hook files to customs subdirectory

* Move store hooks into hooks directory
