+++ src/components/Header/Header.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import { useUserStore } from '~/stores/useUserStore';
-import { useQuotesStore } from '~/stores/useQuotesStore';
-import { useCardStore } from '~/stores/useCardStore';
-import { useModalStore } from '~/stores/useModalStore';
+import { useUserStore } from '~/hooks/stores/useUserStore';
+import { useQuotesStore } from '~/hooks/stores/useQuotesStore';
+import { useCardStore } from '~/hooks/stores/useCardStore';
+import { useModalStore } from '~/hooks/stores/useModalStore';
 import SVGIconBtn from '~/components/common/SVGIconBtn/SVGIconBtn';
 import * as colors from '~/styles/common/colors';
 import * as sizes from '~/styles/common/sizes';
