+++ src/components/LoginModal/LoginModal.tsx
@@ -1,5 +1,5 @@
 import React, { useCallback } from 'react';
-import { useModalStore } from '~/stores/useModalStore';
+import { useModalStore } from '~/hooks/stores/useModalStore';
 import SocialWrapper from './SocialWrapper';
 import { StyledLoginModal } from './LoginModal.styles';
 import * as colors from '~/styles/common/colors';
