+++ src/components/Main/Main.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import Nav from '~/components/Nav/Nav';
 import Card from './Card/Card';
 import ChevronBox from './ChevronBox/ChevronBox';
-import { useCardStore } from '~/stores/useCardStore';
-import { useModalStore } from '~/stores/useModalStore';
+import { useCardStore } from '~/hooks/stores/useCardStore';
+import { useModalStore } from '~/hooks/stores/useModalStore';
 import { CardWrapper, StyledMain } from './Main.styles';
 import { CardPositionType } from '~/types/user.type';
 import LoginModal from '~/components/LoginModal/LoginModal';
